requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2022.6.15
charset-normalizer==2.1.0
flake8==4.0.1
flake8-broken-line==0.4.0
flake8-isort==4.1.1
flake8-plugin-utils==1.3.2
flake8-return==1.1.3
greenlet==1.1.2
idna==3.3
isort==5.10.1
mccabe==0.6.1
pep8-naming==0.13.0
pycodestyle==2.8.0
pyflakes==2.4.0
python-dotenv==0.20.0
python-telegram-bot==13.13
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.39
testfixtures==6.18.5
tornado==6.2
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.10
